+++ extensions/ip-geolocation/CHANGELOG.md
@@ -1,5 +1,9 @@
 # IP Geolocation Changelog
 
+## [Update command name] - 2022-05-26
+
+- Update command name
+
 ## [Initial Version] - 2022-05-15
 
-- Show local and public IP, search geolocation for any IP address or domain.
+- Show local and public IP, lookup geolocation for any IP address or domain.